+++ src/InvestmentApp.Application/Common/Interfaces/IAiChatServiceFactory.cs
@@ -1,0 +1,6 @@
+namespace InvestmentApp.Application.Common.Interfaces;
+
+public interface IAiChatServiceFactory
+{
+    IAiChatService GetService(string provider);
+}